requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
openai==1.3.5
google-generativeai==0.3.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.9
langchain-text-splitters==0.0.1
faiss-cpu==1.13.2
sentence-transformers==2.2.2
pymupdf==1.27.2
python-pptx==0.6.21
openpyxl==3.1.5
python-docx==0.8.11
PyYAML==6.0.1
sqlite-utils==3.35.0
numpy>=2.0.0
aiofiles==23.2.1
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.12.1
email-validator==2.1.0
python-multipart==0.0.6
cors==1.0.1
azure-storage-blob==12.18.3
office365-rest-python-client==2.5.5
